requirements.txt
# Core Python dependencies
pyyaml>=6.0

# Web Framework
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Database
sqlalchemy>=2.0.0

# Auth & Security
pyjwt>=2.8.0
cryptography>=41.0.0

# Templates & Config
jinja2>=3.1.0
python-dotenv>=1.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# IRC Bot
irc>=20.0.0

# Validation
pydantic>=2.5.0

# Background Tasks & Scheduling
apscheduler>=3.10.0

# Web Server
gunicorn>=21.0.0

# Async support
eventlet>=0.33.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Voice automation dependencies
# Wake word detection
openwakeword>=0.5.0
# Alternative: pvporcupine for Porcupine

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0

# Speech recognition
openai-whisper>=20231117

# Text-to-speech
pyttsx3>=2.90

# LLM for intent parsing (optional)
# ollama>=0.1.0  # Requires Ollama server running locally
